+++ crates/ruff/src/rules/flake8_pytest_style/rules/parametrize.rs
@@ -9,7 +9,7 @@ use ruff_python_ast::node::AstNode;
 use ruff_python_ast::parenthesize::parenthesized_range;
 use ruff_python_ast::{self as ast, Arguments, Constant, Decorator, Expr, ExprContext};
 use ruff_python_codegen::Generator;
-use ruff_python_index::CommentRanges;
+use ruff_python_trivia::CommentRanges;
 use ruff_python_trivia::{SimpleTokenKind, SimpleTokenizer};
 use ruff_text_size::{Ranged, TextRange, TextSize};
