+++ crates/ruff/src/rules/pycodestyle/helpers.rs
@@ -3,7 +3,7 @@ use unicode_width::UnicodeWidthStr;
 use ruff_python_ast::node::AnyNodeRef;
 use ruff_python_ast::parenthesize::parenthesized_range;
 use ruff_python_ast::{CmpOp, Expr};
-use ruff_python_index::CommentRanges;
+use ruff_python_trivia::CommentRanges;
 use ruff_source_file::{Line, Locator};
 use ruff_text_size::{Ranged, TextLen, TextRange};
