+++ crates/ruff_python_ast/tests/parenthesize.rs
@@ -1,5 +1,6 @@
 use ruff_python_ast::parenthesize::parenthesized_range;
 use ruff_python_parser::parse_expression;
+use ruff_python_trivia::CommentRanges;
 use ruff_text_size::TextRange;
 
 #[test]
@@ -10,7 +11,12 @@ fn test_parenthesized_name() {
     let bin_op = expr.as_bin_op_expr().unwrap();
     let name = bin_op.left.as_ref();
 
-    let parenthesized = parenthesized_range(name.into(), bin_op.into(), &[], source_code);
+    let parenthesized = parenthesized_range(
+        name.into(),
+        bin_op.into(),
+        &CommentRanges::default(),
+        source_code,
+    );
     assert_eq!(parenthesized, Some(TextRange::new(0.into(), 3.into())));
 }
 
@@ -22,7 +28,12 @@ fn test_non_parenthesized_name() {
     let bin_op = expr.as_bin_op_expr().unwrap();
     let name = bin_op.left.as_ref();
 
-    let parenthesized = parenthesized_range(name.into(), bin_op.into(), &[], source_code);
+    let parenthesized = parenthesized_range(
+        name.into(),
+        bin_op.into(),
+        &CommentRanges::default(),
+        source_code,
+    );
     assert_eq!(parenthesized, None);
 }
 
@@ -35,7 +46,12 @@ fn test_parenthesized_argument() {
     let arguments = &call.arguments;
     let argument = arguments.args.first().unwrap();
 
-    let parenthesized = parenthesized_range(argument.into(), arguments.into(), &[], source_code);
+    let parenthesized = parenthesized_range(
+        argument.into(),
+        arguments.into(),
+        &CommentRanges::default(),
+        source_code,
+    );
     assert_eq!(parenthesized, Some(TextRange::new(2.into(), 5.into())));
 }
 
@@ -48,7 +64,12 @@ fn test_non_parenthesized_argument() {
     let arguments = &call.arguments;
     let argument = arguments.args.first().unwrap();
 
-    let parenthesized = parenthesized_range(argument.into(), arguments.into(), &[], source_code);
+    let parenthesized = parenthesized_range(
+        argument.into(),
+        arguments.into(),
+        &CommentRanges::default(),
+        source_code,
+    );
     assert_eq!(parenthesized, None);
 }
 
@@ -60,7 +81,12 @@ fn test_parenthesized_tuple_member() {
     let tuple = expr.as_tuple_expr().unwrap();
     let member = tuple.elts.last().unwrap();
 
-    let parenthesized = parenthesized_range(member.into(), tuple.into(), &[], source_code);
+    let parenthesized = parenthesized_range(
+        member.into(),
+        tuple.into(),
+        &CommentRanges::default(),
+        source_code,
+    );
     assert_eq!(parenthesized, Some(TextRange::new(4.into(), 7.into())));
 }
 
@@ -72,7 +98,12 @@ fn test_non_parenthesized_tuple_member() {
     let tuple = expr.as_tuple_expr().unwrap();
     let member = tuple.elts.last().unwrap();
 
-    let parenthesized = parenthesized_range(member.into(), tuple.into(), &[], source_code);
+    let parenthesized = parenthesized_range(
+        member.into(),
+        tuple.into(),
+        &CommentRanges::default(),
+        source_code,
+    );
     assert_eq!(parenthesized, None);
 }
 
@@ -84,7 +115,12 @@ fn test_twice_parenthesized_name() {
     let bin_op = expr.as_bin_op_expr().unwrap();
     let name = bin_op.left.as_ref();
 
-    let parenthesized = parenthesized_range(name.into(), bin_op.into(), &[], source_code);
+    let parenthesized = parenthesized_range(
+        name.into(),
+        bin_op.into(),
+        &CommentRanges::default(),
+        source_code,
+    );
     assert_eq!(parenthesized, Some(TextRange::new(0.into(), 5.into())));
 }
 
@@ -97,6 +133,11 @@ fn test_twice_parenthesized_argument() {
     let arguments = &call.arguments;
     let argument = arguments.args.first().unwrap();
 
-    let parenthesized = parenthesized_range(argument.into(), arguments.into(), &[], source_code);
+    let parenthesized = parenthesized_range(
+        argument.into(),
+        arguments.into(),
+        &CommentRanges::default(),
+        source_code,
+    );
     assert_eq!(parenthesized, Some(TextRange::new(2.into(), 11.into())));
 }
