+++ crates/ruff_python_formatter/src/comments/debug.rs
@@ -182,7 +182,7 @@ mod tests {
     use ruff_formatter::SourceCode;
     use ruff_python_ast::node::AnyNode;
     use ruff_python_ast::{StmtBreak, StmtContinue};
-    use ruff_python_index::CommentRanges;
+    use ruff_python_trivia::CommentRanges;
     use ruff_text_size::{TextRange, TextSize};
 
     use crate::comments::map::MultiMap;
