+++ crates/ruff_python_trivia/Cargo.toml
@@ -16,6 +16,7 @@ license = { workspace = true }
 ruff_text_size = { path = "../ruff_text_size" }
 ruff_source_file = { path = "../ruff_source_file" }
 
+itertools = { workspace = true }
 smallvec = { workspace = true }
 unicode-ident = { workspace = true }
