+++ crates/ruff_python_trivia/src/lib.rs
@@ -1,8 +1,10 @@
+mod comment_ranges;
 mod cursor;
 pub mod textwrap;
 mod tokenizer;
 mod whitespace;
 
+pub use comment_ranges::CommentRanges;
 pub use cursor::*;
 pub use tokenizer::*;
 pub use whitespace::*;
