+++ crates/ruff_wasm/Cargo.toml
@@ -28,6 +28,7 @@ ruff_python_index = { path = "../ruff_python_index" }
 ruff_python_parser = { path = "../ruff_python_parser" }
 ruff_source_file = { path = "../ruff_source_file" }
 ruff_text_size = { path = "../ruff_text_size" }
+ruff_python_trivia = { path = "../ruff_python_trivia" }
 ruff_workspace = { path = "../ruff_workspace" }
 
 console_error_panic_hook = { version = "0.1.7", optional = true }
